+++ strings/case_test.go
@@ -11,6 +11,7 @@ func TestToSnake(t *testing.T) {
 		"foo":       "foo",
 		"FooBar":    "foo_bar",
 		"fooBar":    "foo_bar",
+		"Foo_Bar":   "foo_bar",
 		"Foo-Bar":   "foo_bar",
 		"Foo Bar":   "foo_bar",
 		"FOOBar":    "foo_bar",
@@ -22,6 +23,48 @@ func TestToSnake(t *testing.T) {
 	for in, want := range data {
 		if got := strings.ToSnake(in); got != want {
 			t.Errorf("converting '%s' to snake case failed, expected: %s, actual: %s", in, want, got)
+		}
+	}
+}
+
+func TestToSpinal(t *testing.T) {
+	data := map[string]string{
+		"foo":       "foo",
+		"FooBar":    "foo-bar",
+		"fooBar":    "foo-bar",
+		"Foo_Bar":   "foo-bar",
+		"Foo-Bar":   "foo-bar",
+		"Foo Bar":   "foo-bar",
+		"FOOBar":    "foo-bar",
+		"FOOBarBaz": "foo-bar-baz",
+		"FOOBarBAZ": "foo-bar-baz",
+		"Foo_-Bar":  "foo--bar",
+	}
+
+	for in, want := range data {
+		if got := strings.ToSpinal(in); got != want {
+			t.Errorf("converting '%s' to spinal case failed, expected: %s, actual: %s", in, want, got)
+		}
+	}
+}
+
+func TestToTrain(t *testing.T) {
+	data := map[string]string{
+		"foo":       "Foo",
+		"FooBar":    "Foo-Bar",
+		"fooBar":    "Foo-Bar",
+		"Foo_Bar":   "Foo-Bar",
+		"Foo-Bar":   "Foo-Bar",
+		"Foo Bar":   "Foo-Bar",
+		"FOOBar":    "Foo-Bar",
+		"FOOBarBaz": "Foo-Bar-Baz",
+		"FOOBarBAZ": "Foo-Bar-Baz",
+		"Foo_-Bar":  "Foo--Bar",
+	}
+
+	for in, want := range data {
+		if got := strings.ToTrain(in); got != want {
+			t.Errorf("converting '%s' to train case failed, expected: %s, actual: %s", in, want, got)
 		}
 	}
 }